requirements.txt
asgiref==3.4.1
Django==4.1.3
djangorestframework==3.13.1
Pillow==8.2.0
drf-yasg==1.20.0
djangorestframework-simplejwt==5.0.0
